添加留言模块
添加留言模块

1.3

1.2、整合微信支付

修改配置文件、删除误传的jar

1.1

暂时整合App端

1.0

+++ eladmin-app/src/main/java/me/zhengjie/modules/system/service/ArticleContentService.java
@@ -53,7 +53,7 @@ public interface ArticleContentService {
      * @param id ID
      * @return ArticleBodyDto
      */
-    ArticleBodyDto findById(Integer id);
+    ArticleBodyDto findById(Long id);
 
     /**
     * 创建
@@ -71,7 +71,7 @@ public interface ArticleContentService {
     * 多选删除
     * @param ids /
     */
-    void deleteAll(Integer[] ids);
+    void deleteAll(Long[] ids);
 
     /**
     * 导出数据
