+++ eladmin-system/src/main/java/me/zhengjie/modules/system/service/ArticleBodyService.java
@@ -53,7 +53,7 @@ public interface ArticleBodyService {
      * @param id ID
      * @return ArticleBodyDto
      */
-    ArticleBodyDto findById(Integer id);
+    ArticleBodyDto findById(Long id);
 
     /**
     * 创建
@@ -71,7 +71,7 @@ public interface ArticleBodyService {
     * 多选删除
     * @param ids /
     */
-    void deleteAll(Integer[] ids);
+    void deleteAll(Long[] ids);
 
     /**
     * 导出数据
