+++ eladmin-system/src/main/java/me/zhengjie/modules/system/service/SectionService.java
@@ -53,7 +53,7 @@ public interface SectionService {
      * @param id ID
      * @return SectionDto
      */
-    SectionDto findById(Integer id);
+    SectionDto findById(Long id);
 
     /**
     * 创建
@@ -71,7 +71,7 @@ public interface SectionService {
     * 多选删除
     * @param ids /
     */
-    void deleteAll(Integer[] ids);
+    void deleteAll(Long[] ids);
 
     /**
     * 导出数据
